requirements/dev.txt
-r base.txt
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
ruff==0.1.9